Create a test_helper with the MockApi
