+++ test/game_spec.rb
@@ -1,29 +1,5 @@
 # frozen_string_literal: true
-require 'minitest/autorun'
-require 'mlb_gameday.rb'
-
-# Mock the basic `open` function so we don't actually hit the MLB website
-class MockedApi < MLBGameday::API
-  alias old_open open
-
-  def open(url, &block)
-    dir = File.dirname __FILE__
-    base = url.gsub 'http://gd2.mlb.com/components/game/mlb/', ''
-    path = File.join dir, base
-
-    unless File.exist?(path)
-      puts "Downloading from website: #{url}"
-
-      return old_open(url, &block)
-    end
-
-    file = File.open path
-
-    return file unless block_given?
-
-    yield file
-  end
-end
+require './test/test_helper'
 
 class TestGame < MiniTest::Test
   def setup
